Share Menu. More code refactoring and fixes (#734)
