+++ src/main/java/io/kyberorg/yalsee/ui/elements/shareitem/ShareItem.java
@@ -18,7 +18,7 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 public abstract class ShareItem extends Composite<VerticalLayout> {
-    static final String DEFAULT_SHORT_LINK = "https://yals.ee";
+    protected static final String DEFAULT_SHORT_LINK = "https://yals.ee";
     protected static final String DEFAULT_DESCRIPTION = "Yalsee: Yet another link shortener";
 
     private final Image logo;
@@ -30,17 +30,39 @@ public abstract class ShareItem extends Composite<VerticalLayout> {
     @Getter(value = AccessLevel.PROTECTED)
     private String description = DEFAULT_DESCRIPTION;
 
+    /**
+     * Provides {@link #DEFAULT_SHORT_LINK} to classes that are not {@link ShareItem}.
+     *
+     * @return string with {@link #DEFAULT_SHORT_LINK}
+     */
+    public static String getDefaultShortLink() {
+        return DEFAULT_SHORT_LINK;
+    }
+
+    /**
+     * Updates short link.
+     *
+     * @param newShortLink not-empty string with short link.
+     */
     public void updateShortLink(final String newShortLink) {
         this.shortLink = newShortLink;
         this.description = "";
         constructLink();
     }
 
+    /**
+     * Updates description.
+     *
+     * @param newDescription not-empty string with description
+     */
     public void updateDescription(final String newDescription) {
         this.description = newDescription;
         constructLink();
     }
 
+    /**
+     * Creates {@link ShareItem}.
+     */
     protected ShareItem() {
         logo = new Image();
         logo.setSrc("/images/logo.png");
@@ -52,10 +74,20 @@ public abstract class ShareItem extends Composite<VerticalLayout> {
         getContent().add(logo, label);
     }
 
+    /**
+     * Sets Item's Icon.
+     *
+     * @param icon icon from selections.
+     */
     protected void setImageFile(final ShareMenu.Icons icon) {
         logo.setSrc("/images/" + icon.getFileName());
     }
 
+    /**
+     * Item's Label aka Name.
+     *
+     * @param labelText non-empty string with item's label or name.
+     */
     protected void setLabelText(final String labelText) {
         label.setText(labelText);
     }
@@ -65,6 +97,11 @@ public abstract class ShareItem extends Composite<VerticalLayout> {
      */
     protected abstract void constructLink();
 
+    /**
+     * Action that should happen when user clicks on icon.
+     *
+     * @param clickEvent Vaadin's {@link ClickEvent}, populated by Vaadin itself.
+     */
     protected void onLogoClick(final ClickEvent<Image> clickEvent) {
         UI ui = null;
         if (getUI().isPresent()) {
