+++ src/main/java/io/kyberorg/yalsee/ui/pages/home/HomePage.java
@@ -35,8 +35,8 @@ import io.kyberorg.yalsee.services.QRCodeService;
 import io.kyberorg.yalsee.services.overall.OverallService;
 import io.kyberorg.yalsee.session.YalseeSession;
 import io.kyberorg.yalsee.ui.MainView;
+import io.kyberorg.yalsee.ui.elements.MobileShareMenu;
 import io.kyberorg.yalsee.ui.elements.ShareMenu;
-import io.kyberorg.yalsee.ui.elements.shareitem.MobileShareMenu;
 import io.kyberorg.yalsee.utils.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
