+++ src/main/java/io/kyberorg/yalsee/ui/pages/mylinks/MyLinksPage.java
@@ -49,8 +49,8 @@ import io.kyberorg.yalsee.services.YalseeSessionService;
 import io.kyberorg.yalsee.session.YalseeSession;
 import io.kyberorg.yalsee.ui.MainView;
 import io.kyberorg.yalsee.ui.elements.DeleteConfirmationDialog;
+import io.kyberorg.yalsee.ui.elements.MobileShareMenu;
 import io.kyberorg.yalsee.ui.elements.ShareMenu;
-import io.kyberorg.yalsee.ui.elements.shareitem.MobileShareMenu;
 import io.kyberorg.yalsee.ui.layouts.YalseeLayout;
 import io.kyberorg.yalsee.utils.AppUtils;
 import io.kyberorg.yalsee.utils.ClipboardUtils;
